+++ src/__tests__/components/FeedList/FeedList.test.tsx
@@ -1,10 +1,10 @@
 import { mount, shallow } from 'enzyme'
 import FeedList from 'components/FeedList/FeedList'
-import { Feed } from 'domain/Feed'
 import useSmallScreen from 'hooks/useSmallScreen'
 import React, { useEffect } from 'react'
 import FeedItem, { FeedItemProps } from 'components/FeedItem/FeedItem'
 import { act } from 'react-dom/test-utils'
+import { Feed } from 'shared/api'
 
 jest.mock(
   'shared/ui/Loader',
