--- src/domain/AuthorMeta.ts
@@ -1,9 +1,0 @@
-export interface AuthorMeta {
-  id: string
-  name: string
-  avatar: string
-  heart: number
-  video: number
-  fans: number
-  following: number
-}
