+++ src/shared/api/models.ts
@@ -1,0 +1,2 @@
+export * from './feed/models'
+export * from './user/models'
